[hitbtc] Fix lot multipliers for currencies, change order quantity to BigInteger

+++ xchange-hitbtc/src/main/java/com/xeiam/xchange/hitbtc/HitbtcAuthenticated.java
@@ -2,6 +2,7 @@ package com.xeiam.xchange.hitbtc;
 
 import java.io.IOException;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.FormParam;
@@ -38,7 +39,7 @@ public interface HitbtcAuthenticated extends Hitbtc {
   @Path("trading/new_order")
   public HitbtcExecutionReportResponse postHitbtcNewOrder(@HeaderParam("X-Signature") ParamsDigest signature, @QueryParam("nonce") SynchronizedValueFactory<Long> valueFactory, @QueryParam("apikey") String apiKey,
       @FormParam("clientOrderId") String clientOrderId, @FormParam("symbol") String symbol, @FormParam("side") String side, @FormParam("price") BigDecimal price,
-      @FormParam("quantity") BigDecimal quantity, // 1 lot = 0.01 BTC
+      @FormParam("quantity") BigInteger quantity, // 1 lot = 0.01 BTC
       @FormParam("type") String type, @FormParam("timeInForce") String timeInForce) throws IOException;
 
   @POST
